requirements.txt
pytorch-ignite==0.4.2
transformers==4.8.2
tensorboardX==2.1
tqdm==4.60.0
protobuf==3.20.1